refactor: create @hyperframes/lint package
