+++ packages/core/package.json
@@ -281,6 +281,7 @@
   },
   "dependencies": {
     "@chenglou/pretext": "^0.0.5",
+    "@hyperframes/lint": "workspace:*",
     "@hyperframes/parsers": "workspace:*",
     "bpm-detective": "^2.0.5",
     "postcss": "^8.5.8",
