+++ packages/core/src/slideshow/index.ts
@@ -1,2 +1,3 @@
 export * from "./slideshow.types";
 export * from "./parseSlideshow";
+export { isSceneLikeCompositionId } from "./sceneId";
