+++ packages/lint/src/rules/adapters.test.ts
