+++ packages/lint/src/rules/adapters.ts
