+++ packages/lint/src/rules/captions.test.ts
