+++ packages/lint/src/rules/captions.ts
