+++ packages/lint/src/rules/core.test.ts
