+++ packages/lint/src/rules/core.ts
