+++ packages/lint/src/rules/gsap.test.ts
