+++ packages/lint/src/rules/slideshow.test.ts
