+++ packages/lint/src/rules/textures.test.ts
